feat: Create a framework for managing this extension's database schema
MW core's DatabaseUpdater has stuff in it to manage DB bits for an
extension --- but it is kind of too smart and not smart enough at
the same time.  I.e., I was trying to figure out how to add a
unique constraint/index to a column, behave nicely if the patch to
do that failed (e.g., due to a constraint violation), ensure that
the patch is applied only once, and ensure that the patch has been
applied at all --- and I didn't see a way with the existing machinery.

So, I cooked up something that keeps track of a simple monotonically
increasing patch-level ("schema version"), with idempotent patch files
that can only be successfully applied once and must be applied in order.
The PHP code is interlocked with the DB schema, and will refuse to run
if the DB's schema version does not match what the PHP code expects.