fix:  get Discourse logout working again, by using a log_out API
The MediaWiki logout flow changed between 1.31 and 1.35.  Pressing the
logout button used to be a simple navigation/GET request, so returning
a browser redirect to the Discourse SSO endpoint (with a logout request)
worked.  Now, the standard logout flow invokes a javascript POST to
`api.php`, staying on the same page (until the result comes back, etc).
The original redirect-to-SSO-to-logout method might still work, but it
has to deal with CORS issues (since it is no longer a navigation request),
etc, etc, ... it makes more sense to have MediaWiki go and make its own
side-band request to the Discourse `log_out` API endpoint, especially
since the user (and their cookies) do not need to be in-the-loop for a
logout.

Fixes #3.