fix: Correct the schema-v0008 files to identify schema as v8, not v7.
The install-entire-schema-from-scratch SQL files for schema version 8,
used for brand-new installation of the extension, mistakenly identified
the resulting schema as version 7 (a classic cut'n'paste error).

As a result, with new installs the extension would complain that the
database needed to be upgraded to version 8, and trying to run the
MW upgrade script would fail with a database error (trying to create
a table that already existed).

The patch files used for incremental upgrades did not have this problem,
and incremental upgrades were not affected.